backend/app/core/jwt.py
from datetime import datetime, timedelta
from jose import jwt
from fastapi import HTTPException, status

SECRET_KEY = "63f4945d921d599f27ae4fdf5bada3f1"
ALGORITHM = "HS256"

def create_access_token(user_id: int, role: str = "admin", expires_minutes: int = 60):
    expire = datetime.utcnow() + timedelta(minutes=expires_minutes)
    payload = {
        "sub": str(user_id),
        "role": role,
        "exp": expire
    }
    token = jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
    return token


def decode_access_token(token: str):
    try:
        payload = jwt.decode(
            token,
            SECRET_KEY,
            algorithms=[ALGORITHM]
        )
        return payload
    except jwt.JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token"
        )
