backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.23.2
SQLAlchemy>=2.0
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib[bcrypt]
pydantic==2.7.2
